apis/requirements.txt
openai
llama-index
PyPDF2
guardrails-ai
langchain
kor
Werkzeug==2.2.2
flask
python-dotenv
jsonify